+++ nd4j-api/src/main/java/org/nd4j/linalg/ops/transforms/GreaterThan.java
@@ -22,14 +22,14 @@ public class GreaterThan extends BaseElementWiseOp {
     @Override
     public Object apply(INDArray from,Object value, int i) {
         Object curr = getFromOrigin(from,i);
-        float originValue = (float) (curr instanceof IComplexNDArray ? ((IComplexNumber) curr).absoluteValue() : (float) curr);
-        float otherValue = (float) (value instanceof IComplexNDArray ? ((IComplexNumber) value).absoluteValue() : (float) value);
+        double originValue = (double) (curr instanceof IComplexNDArray ? ((IComplexNumber) curr).absoluteValue() : (double) curr);
+        double otherValue = (double) (value instanceof IComplexNDArray ? ((IComplexNumber) value).absoluteValue() : (double) value);
         if(originValue > otherValue) {
             if(value instanceof IComplexNumber) {
                 return  Nd4j.createDouble(1, 0);
             }
             else
-                return (float)  1;
+                return (double)  1;
 
 
         }
@@ -38,7 +38,7 @@ public class GreaterThan extends BaseElementWiseOp {
                 return Nd4j.createDouble(0, 0);
 
             else
-                return (float) 0;
+                return (double) 0;
 
 
         }
