+++ nd4j-api/src/main/java/org/nd4j/linalg/ops/transforms/LessThan.java
@@ -22,8 +22,8 @@ public class LessThan extends BaseElementWiseOp {
     @Override
     public Object apply(INDArray from,Object value, int i) {
         Object curr = getFromOrigin(from,i);
-        float originValue = (float) (curr instanceof IComplexNDArray ? ((IComplexNumber) curr).absoluteValue() : (float) curr);
-        float otherValue = (float) (value instanceof IComplexNDArray ? ((IComplexNumber) value).absoluteValue() : (float) value);
+        double originValue = (double) (curr instanceof IComplexNDArray ? ((IComplexNumber) curr).absoluteValue() : (double) curr);
+        double otherValue = (double) (value instanceof IComplexNDArray ? ((IComplexNumber) value).absoluteValue() : (double) value);
         if(originValue < otherValue) {
             if(value instanceof IComplexNumber)
                 return Nd4j.createDouble(1, 0);
