+++ nd4j-api/src/main/java/org/nd4j/linalg/ops/transforms/Tanh.java
@@ -27,8 +27,8 @@ public class Tanh extends BaseElementWiseOp {
             return ComplexUtil.tanh(element);
         }
         else  {
-            float d = (float) value;
-            return  (float)  FastMath.tanh(d);
+            double d = (double) value;
+            return   FastMath.tanh(d);
         }
     }
 }
